Removing prop checker from default middleware, refactored middleware
