+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "redux-modules",
-  "version": "0.5.3",
+  "version": "0.6.0",
   "description": "A library for defining clear, boilerplate free Redux reducers.",
   "main": "lib/index.js",
   "scripts": {
