+++ src/actionMiddleware/parsePayloadErrors.js
@@ -1,0 +1,7 @@
+const parsePayloadErrors = (transformation, { payload, meta }) => ({
+  payload,
+  meta,
+  error: (payload instanceof Error),
+});
+
+export default parsePayloadErrors;
