+++ src/createModule/createAction.js
@@ -2,8 +2,10 @@ const createAction = (transform, defaultMiddleware = []) => {
   const { formattedConstant, middleware = [] } = transform;
   const actionMiddleware = middleware.concat(defaultMiddleware);
   return (payload, meta) => ({
-    type: formattedConstant,
-    ...actionMiddleware.reduce((acc, func) => func(transform, acc), { payload, meta }),
+    ...actionMiddleware.reduce(
+      (acc, func) => func(acc),
+      { payload, meta, type: formattedConstant }
+    ),
   });
 };
